# resources/swagger/georchestra-signalement.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Georchestra - Plugin Signalement API",
		"contact": {
			"email": "guillaume.morin@open-groupe.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "",
	"basePath": "/signalement",
...
	"schemes": [
		"https",
		"http"
	],
	"paths": {
...
		"/reporting/attachment/configuration": {
			"get": {
				"tags": [
					"reporting"
				],
				"summary": "Lister la configuration liées aux pièces jointes",
				"description": "Lister la configuration liées aux pièces jointes",
				"operationId": "getAttachmentConfiguration",
				"produces": [
					"application/json"
				],
				"parameters": [],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/AttachmentConfiguration"
						}
					},
					"400": {
						"description": "bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/reporting/contextDescription/search": {
			"get": {
				"tags": [
					"reporting",
					"contextDescriptions"
				],
				"summary": "Lister les contextes",
				"description": "Listes les contexts selon le type",
				"operationId": "searchReportingContextDescriptions",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "contextType",
						"description": "Type du context",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "geographicType",
						"description": "Type de sélection",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/ContextDescription"
							}
						}
					},
					"400": {
						"description": "bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/reporting/form/draft/{contextDescriptionName}": {
			"get": {
				"tags": [
					"reporting"
				],
				"summary": "Retourne le formulaire draft",
				"description": "Retourne le formulaire associé à l'état draft",
				"operationId": "lookupDrafForm",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "contextDescriptionName",
						"description": "Le nom du contexte, thème ou couche",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Form"
						}
					},
					"400": {
						"description": "bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/reporting/{uuid}/upload": {
			"post": {
				"tags": [
					"reporting"
				],
				"summary": "Téléversement d'une pièce jointe sur un signalement",
				"description": "",
				"operationId": "uploadDocument",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"description": "Identifiant du signalement",
						"required": true,
						"type": "string",
						"format": "uuid"
					},
					{
						"in": "formData",
						"name": "file",
						"description": "Pièce jointe sous la forme d'un blob",
						"required": true,
						"type": "file"
					}
				],
				"consumes": [
					"multipart/form-data",
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Attachment"
						}
					},
					"400": {
						"description": "bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/reporting/{uuid}/delete/{attachmentId}": {
			"delete": {
				"tags": [
					"reporting"
				],
				"summary": "Suppression d'un pièce jointe sur un signalement",
				"description": "",
				"operationId": "deleteDocument",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"description": "Identifiant du signalement",
						"required": true,
						"type": "string",
						"format": "uuid"
					},
					{
						"in": "path",
						"name": "attachmentId",
						"description": "Identifiant de l'attachement",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/reporting/{uuid}/download/{attachmentId}": {
			"get": {
				"tags": [
					"reporting"
				],
				"summary": "Téléchargement d'un pièce attachée à un signalement",
				"description": "",
				"operationId": "downloadDocument",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"description": "Identifiant du signalement",
						"required": true,
						"type": "string",
						"format": "uuid"
					},
					{
						"in": "path",
						"name": "attachmentId",
						"description": "Identifiant de l'attachement",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "file"
						}
					},
					"400": {
						"description": "bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
...
		"/contextDescriptions": {
			"get": {
				"tags": [
					"contextDescriptions"
				],
				"summary": "Liste les desctiptions de contexte disponibles.",
				"description": "Retourne la liste des desctiptions de contexte existantes",
				"operationId": "searchContextDescriptions",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "description",
						"description": "Filtre de recherche sur la description",
						"required": false,
						"type": "string",
						"default": ""
					},
					{
						"in": "query",
						"name": "workflow",
						"description": "Filtre de recherche sur le workflow",
						"required": false,
						"type": "string",
						"default": ""
					},
					{
						"in": "query",
						"name": "offset",
						"description": "Index de début",
						"required": false,
						"type": "integer"
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Nombre de résultats à retourner",
						"required": false,
						"type": "integer"
					},
					{
						"in": "query",
						"name": "sortExpression",
						"description": "Indications de tri",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/ContextDescriptionPageResult"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"post": {
				"tags": [
					"contextDescriptions"
				],
				"summary": "Créer une desctiption de contexte.",
				"description": "Créer une desctiption de contexte telle que stipulée dans le body de la requête",
				"operationId": "createContextDescription",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "contextDescription",
						"description": "Description du contexte",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ContextDescription"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/ContextDescription"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"put": {
				"tags": [
					"contextDescriptions"
				],
				"summary": "Mise à jour d'une description de contexte.",
				"description": "Mise à jour d'une desctiption de contexte telle que stipulée dans le body de la requête, en fonction du nom de la description",
				"operationId": "updateContextDescription",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "contextDescription",
						"description": "Description du contexte",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ContextDescription"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/ContextDescription"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
...
		"/roles": {
			"get": {
				"tags": [
					"roles"
				],
				"summary": "Récupération de la liste des rôles.",
				"description": "Récupération de la liste des rôles disponibles.",
				"operationId": "searchRoles",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"description": "Index de début",
						"required": false,
						"type": "integer"
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Nombre de résultats à retourner",
						"required": false,
						"type": "integer"
					},
					{
						"in": "query",
						"name": "sortExpression",
						"description": "Indications de tri",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/RolePageResult"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"post": {
				"tags": [
					"roles"
				],
				"summary": "Création d'un rôle.",
				"description": "Création d'un rôle.",
				"operationId": "createRole",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "role",
						"description": "Rôle à créer",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Role"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/Role"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/roles/{name}": {
			"get": {
				"tags": [
					"roles"
				],
				"summary": "Récupération d'un rôle.",
				"description": "Récupération d'un rôle en fonction de son nom",
				"operationId": "getRole",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"description": "Nom du rôle à récupérer",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/Role"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"roles"
				],
				"summary": "Suppression d'un rôle.",
				"description": "Suppression d'un rôle en fonction de son nom",
				"operationId": "deleteRole",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"description": "Nom du rôle à supprimer",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
...
	},
	"definitions": {
		"ApiError": {
			"type": "object",
			"properties": {
				"code": {
					"type": "string"
				},
				"label": {
					"type": "string"
				}
			}
		},
...
		"GeographicType": {
			"type": "string",
			"description": "Type géographic",
			"enum": [
				"POINT",
				"LINE",
				"POLYGON"
			]
		},
...
		"ContextType": {
			"type": "string",
			"description": "Type du context du signalement",
			"enum": [
				"THEMA",
				"LAYER"
			]
		},
		"FieldType": {
			"type": "string",
			"description": "Type des champs",
			"enum": [
				"LONG",
				"DOUBLE",
				"STRING",
				"BOOLEAN",
				"DATE",
				"LIST"
			]
		},
		"ValidatorType": {
			"type": "string",
			"enum": [
				"REQUIRED",
				"MAXLENGTH",
				"POSITIVE",
				"NEGATIVE",
				"REGEXP"
			]
		},
		"Validator": {
			"type": "object",
			"properties": {
				"type": {
					"$ref": "#/definitions/ValidatorType"
				},
				"attribute": {
					"type": "string"
				}
			}
		},
		"FieldDefinition": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				},
				"help": {
					"type": "string"
				},
				"type": {
					"type": "object",
					"$ref": "#/definitions/FieldType"
				},
				"readOnly": {
					"type": "boolean"
				},
				"required": {
					"type": "boolean"
				},
				"multiple": {
					"type": "boolean"
				},
				"extendedType": {
					"type": "string"
				},
				"validators": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Validator"
					}
				}
			}
		},
		"Field": {
			"type": "object",
			"properties": {
				"values": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"definition": {
					"type": "object",
					"$ref": "#/definitions/FieldDefinition"
				}
			}
		},
		"Section": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				},
				"readOnly": {
					"type": "boolean"
				},
				"fields": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/Field"
					}
				}
			}
		},
		"Form": {
			"type": "object",
			"properties": {
				"sections": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/Section"
					}
				}
			}
		},
		"ContextDescription": {
			"type": "object",
			"properties": {
				"contextType": {
					"type": "object",
					"$ref": "#/definitions/ContextType"
				},
				"geographicType": {
					"type": "object",
					"$ref": "#/definitions/GeographicType"
				},
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				},
				"processDefinitionKey": {
					"type": "string"
				},
				"revision": {
					"type": "integer"
				}
			}
		},
		"ContextDescriptionPageResult": {
			"type": "object",
			"properties": {
				"totalItems": {
					"type": "integer",
					"format": "int64"
				},
				"results": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/ContextDescription"
					}
				}
			}
		},
...
		"Attachment": {
			"type": "object",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"name": {
					"type": "string"
				},
				"mimeType": {
					"type": "string"
				}
			}
		},
		"AttachmentConfiguration": {
			"type": "object",
			"properties": {
				"maxSize": {
					"type": "integer",
					"format": "int64"
				},
				"maxCount": {
					"type": "integer",
					"format": "int32"
				},
				"mimeTypes": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"mimeTypesByExtension": {
					"type": "object",
					"description": "Mapping des extensions de fichier (sans point, en minuscules) vers leur type MIME autorisé. Calculé depuis la configuration serveur.",
					"additionalProperties": {
						"type": "string"
					}
				}
			}
		},
...
		"Role": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				}
			}
		},
		"RolePageResult": {
			"type": "object",
			"properties": {
				"totalItems": {
					"type": "integer",
					"format": "int64"
				},
				"results": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/Role"
					}
				}
			}
		},
...
	}
}